+++ src/go/types/check.go
@@ -411,7 +411,7 @@ func (check *Checker) recordCommaOkTypes(x ast.Expr, a [2]Type) {
 func (check *Checker) recordInferred(call ast.Expr, targs []Type, sig *Signature) {
 	assert(call != nil)
 	assert(sig != nil)
-	if m := check._Inferred; m != nil {
+	if m := getInferred(check.Info); m != nil {
 		m[call] = _Inferred{targs, sig}
 	}
 }
